check current ip in background
